feat(BANK-9): Cria script para limpar database e para popular com 10 elementos cada tabela
